feat(cypress): adding more tests to the pricing table cells (#11286)
### Related Ticket(s)

Closes # [ADCMS-4051](https://jsw.ibm.com/browse/ADCMS-4051)

### Description

We created a regression when implementing the Pricing Table Cell feature to align text next to the optional icons. Then we almost wiped out the side-by-side alignment feature when fixing the regression. This all occurred because our current tests don't adequately reflect our expected functionality.

We need tests for the following cases:

1.     When the icon feature is enabled, user-entered cell content should appear next to the icon.
2.     User-entered content should appear in a vertical column layout whether there's an icon enabled or not.