fix(tabs-extended): confusing behavior when in mobile the accordion c… (#11595)
…annot be closed

### Related Ticket(s)

[ADCMS-4434](https://jsw.ibm.com/browse/ADCMS-4434)

### Description

The design for the Tabs Extended when in mobile is confusing. It turns into an accordion and there's always one item opened and showing. However, that opened item has an arrow on the right hand side, that deceives the user into thinking it's possible to close that item, when in fact, the accordion mimics the desktop behavior for tabs where we always have one item being shown at a time.

**I'm opening this PR initially as draft with the solution of removing the arrow for opened items so the whole team can discuss the best approach here.**

As per discussion with the team, the suggested changes were:


1.     Added the ability to open and close the accordion items;
2.     When all accordion items are closed, in desktop view the first tab is selected as default.