fix(pricing-table): increasing annotation toggle touch area (#11623)
### Related Ticket(s)

[ADCMS-4432](https://jsw.ibm.com/browse/ADCMS-4432)

### Description

In continuation of [this PR for v1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/11581). I'm raising this one for v2:

In the pricing table, the toggle annotation arrow was too small, making it difficult to be hit on mobile resolutions. I added more padding so it's now easier to be touched.