[v1] feat(rtl-arrow): Fix arrows used for CTA's when in RTL (#11748)
### Related Ticket(s)

Closes https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/11725
[Jira Ticket](https://jsw.ibm.com/browse/ADCMS-4922)

### Description

Checks `document.dir` and uses the appropriate arrow icon for the text direction (right in LTR and left in RTL). For many stories involving CTA's, CTA derivatives, or buttons, avoid using the slotted icon when all we're going for is `cta-type="local"`, in other words the arrow icon. In non-CTA cases, there isn't support for a default arrow icon, so we adjust stories to slot the right icon base on `document.dir`.

Note that in v1, only 'CTA' components have the `CTAMixin`, whereas it's more broadly used in v2. This means that for non-CTA components its up to the client to slot the right arrow icon (as evidenced by the many `*.stories.ts` changes). This isn't great, but it felt OK for now 🤷 . Open to ideas however.

### Testing instructions

1. Navigate to any Story making use of a CTA and/or button
2. Compare LTR and RTL
3. In LTR, where the arrow icon is used, the arrow should be pointing right
4. In RTL, where the arrow icon is used, the arrow should be pointing left

### Changelog

**Changed**

- Change the direction of the CTA arrow when in RTL languages.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->