feat(sticky-header): allow mixed prefixes with warning (#11868)
### Related Ticket(s)

[ADCMS-5224](https://jsw.ibm.com/browse/ADCMS-5224)

### Description

The StickyHeader utility validates all components it tracks, prior to this PR is expects the provided element's `tagName` to match `${c4dPrefix}-component-name`. This works when all components are using the same prefix, but as we work on upgrading from v1 to v2 we'll need to mix-and-match prefixes. This PR will allow the StickyHeader to continue working despite the prefix differences

### Changelog

**Changed**

- allows StickyHeader to validate (with a warning) components that have unexpected prefixes.