fix(tabs-extended): prevent stealing focus on page load (#11882)
### Related Ticket(s)

none

### Description

The current implementation sets focus to a tab any time it is set as "active". This works great for the intended use case, but the problem is that the first tab is made active when the component mounts. This is not so great on page load, forcing focus to the tab and scrolling the page down to it.

The fix is to only set focus on a user interaction.

### Changelog

**Changed**

- Prevents DDSTabsExtended elements from stealing focus on component mount.

### Testing Instructions

1. Visit the Tabs Extended deploy preview and "open the canvas in a new tab" (https://ibmdotcom-webcomponents.s3.us-east.cloud-object-storage.appdomain.cloud/deploy-previews/11882/iframe.html?args=&id=components-tabs-extended--default&viewMode=story)
2. Using your browser console, check which element has focus on load – `document.activeElement`
   - Recommend using Chrome's live expression feature and refreshing the page
3. Verify the focused element is _not_ the Tabs Extended component.

Additionally, you can use the CDN preview on an affected page on IBM.com, such as https://www.ibm.com/products/openpages. Verify the page doesn't scroll down to the Tabs Extended Media component on load.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->