fix(cta-block-item-row): set statistics same height (#11895)
### Related Ticket(s)


### Description

`<dds-cta-block-item-row>` elements use the `sameHeight` utility to equalize the heights of slotted headings and text, but not statistics. This PR adds support for that and adds documentation in the form of a Storybook example.

### Testing

Visit the [CTA Block > With Content Items deploy preview story](https://ibmdotcom-webcomponents.s3.us-east.cloud-object-storage.appdomain.cloud/deploy-previews/11895/index.html?path=/story/components-cta-block--with-content-items) and select the "Show statistics?" checkbox. Open the canvas in a new tab (or resize your browser window) and verify the statistics are given the same height.

### Changelog

**Changed**

- Enables `<dds-cta-block-item>`s' slotted statistics to have their height equalized when in a `<dds-cta-block-item-row>`.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->