fix(text-cta): prevent order of operations issues (#11911)
### Related Ticket(s)
[Jira](https://jsw.ibm.com/browse/ADCMS-5207)
### Description
An issue where the CTA on cards with videos is not being populated on some pages is being caused  by an order of operation issue where the element is making calls tho an event listener that is not yet defined. 
This PR is to ensure the correct order of execution  

### Changelog

**Changed**
Added `customElements.whenDefined()` to the video mixin  to make sure the event only triggers after all componets needed are loaded.