fix(carousel): Fix gap calculation (#12179)
### Related Ticket(s)

[JIRA](https://jsw.ibm.com/browse/ADCMS-7404)
### Description

PR To address the carousel component gap distance not being calculated properly. causing the slotted components to scroll too much and appear cropped.

### Changelog

**Changed**

changed the `_updateGap` function to calculate the gap value based on the margin of the fist element. 

### Testing instructions 
Open the carousel components on storybook and see that it scrolls properly and there are no regressions.