feat(feature-section): adding media alignment (#12195)
### Related Ticket(s)

Closes #[ADCMS-7463](https://jsw.ibm.com/browse/ADCMS-7463)

### Description

On CIBM v1, the Feature Card Section component allowed the author to switch the positioning of the media between the right and left sides on the component. This capability was removed in v2 and the UX team is recommending to bring this capability back.

### Changelog

**New**
Media alignment has been added back to the `feature-section` component.