fix(toc): centering clicked tabs in mobile (#12209)
### Related Ticket(s)

Closes [#6635](https://jsw.ibm.com/browse/ADCMS-6635)

### Description

The table of contents changes its navigation to a horizontal scrollable tabs section in mobile. When clicking through the tabs, there are tabs that can partially exist off the visible area. When they're clicked, the page scrolls to the desired content but the tab will remain there it is, sometimes being partially off the screen.

### Changelog

I added a `scrollIntoView()` that will make each clicked tab to scroll to the center of the visible area, ensuring its copy is fully visible


### Testing Instructions:

1. In the storybook, go to the `Table of Contents > Default` component and enter mobile view.
2. Click the tabs and they need to scroll to the center of the visible area instead of not moving



https://github.com/user-attachments/assets/e6cd7cc3-3089-4d6f-a9cc-a21513fbcfc3